+++ Reach/test/utopia/reach/test/DragToResizeWindowTest.scala
@@ -6,7 +6,7 @@ import utopia.firmament.model.Border
 import utopia.paradigm.color.Color
 import utopia.paradigm.shape.shape2d.insets.Insets
 import utopia.reach.component.label.empty.EmptyLabel
-import utopia.reach.cursor.{DragTo, DragToResize}
+import utopia.reach.cursor.DragTo
 import utopia.reach.window.ReachWindow
 
 /**