+++ Access/Changes.md
@@ -3,7 +3,10 @@
 ## v1.5 (in development)
 Supports **Flow v2.3**
 ### Breaking changes
+- **Headers**`.apply(String)` and other header-retrieving functions now return a **String** instead of an **Option**
 - `.isTemporary` and `.doNotRepeat` in **Status** are now of type **UncertainBoolean** instead of **Boolean**
+### Bugfixes
+- Header-retrieving functions didn't work at all because of a casing issue - fixed
 ### Other changes
 - Scala version updated to 2.13.12